// frontend/src/client/types.gen.ts
export type ReviewPublic = {
    content: string;
    id: number;
    book_id: number;
    owner_id?: (string | null);
    book_title?: (string | null);
    author_name?: (string | null);
    created_at?: (string | null);
};